Feature: 토큰 인증 및 userId 일치 검사 추가

+++ src/main/java/com/backend/wear/controller/ProductController.java
@@ -4,6 +4,7 @@ import com.backend.wear.dto.product.ProductPostRequestDto;
 import com.backend.wear.dto.product.ProductRequestDto;
 import com.backend.wear.dto.product.ProductResponseDto;
 import com.backend.wear.service.ProductService;
+import com.backend.wear.service.TokenService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -11,6 +12,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
@@ -21,10 +23,12 @@ import java.util.concurrent.CompletableFuture;
 public class ProductController {
 
     private final ProductService productService;
+    private final TokenService tokenService;
 
     @Autowired
-    public ProductController(ProductService productService){
+    public ProductController(ProductService productService, TokenService tokenService){
         this.productService=productService;
+        this.tokenService=tokenService;
     }
 
     // 카테고리별 최신순 조회
@@ -57,23 +61,6 @@ public class ProductController {
                     .body(e.getMessage());
         }
     }
-
-    //검색어 입력 후 검색어별, 최신순(default)으로 조회하기
-    //  /products/search?searchName={searchName}&userId={userId}&pageNumber={pageNumber}
-//    @GetMapping("/search")
-//    public ResponseEntity<?> searchProducts(@RequestParam(name="searchName") String searchName,@RequestParam(name="userId") Long userId,
-//                                            @RequestParam(name="pageNumber")Integer pageNumber) throws Exception {
-//
-//        try {
-//            Page <ProductResponseInnerDto.ScreenDto> productsPage
-//                    = productService.searchProductByproductName(searchName, userId, pageNumber);
-//            return ResponseEntity.ok(productsPage);
-//        } catch (IllegalArgumentException e) {
-//            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-//                    .body(e.getMessage());
-//        }
-//
-//    }
 
     //상품 리스트 검색어별, 카테고리별 , 최신순(default)으로 조회하기
     //  /products/search/category?searchName={searchName}&userId={userId}&pageNumber={pageNumber}
@@ -114,8 +101,12 @@ public class ProductController {
     //상품 상세 페이지 불러오기
     // api/products/{productId}
     @GetMapping("/{userId}/{productId}")
-    public ResponseEntity<?> getProductPost(@PathVariable("userId") Long userId,@PathVariable("productId") Long productId) throws Exception {
-        ;
+    public ResponseEntity<?> getProductPost(@PathVariable("userId") Long userId,@PathVariable("productId") Long productId,
+                                            @RequestHeader("Authorization") String authorizationHeader) throws Exception {
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             ProductResponseDto.DetailDto productPost = productService.getProductPost(userId, productId);
             return ResponseEntity.ok(productPost);
@@ -127,8 +118,13 @@ public class ProductController {
 
     // 상품 등록
     @PostMapping("/new/{userId}")
-    public ResponseEntity<?> postProductPost(@PathVariable(name="userId") Long userId , @RequestBody @Valid ProductPostRequestDto requestDTO, Errors errors)
-            throws Exception{
+    public ResponseEntity<?> postProductPost(@PathVariable(name="userId") Long userId , @RequestBody @Valid ProductPostRequestDto requestDTO,
+                                             @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.createProductPost(requestDTO,userId);
         } catch (IllegalArgumentException e) {
@@ -141,8 +137,12 @@ public class ProductController {
 
     // 상품 상세/상품 정보 수정하기로 이동
     @GetMapping("/edit/{userId}/{productId}")
-    public ResponseEntity<?> getProductPostToUpdate(@PathVariable(name="userId") Long userId , @PathVariable(name="productId") Long productId)
-            throws Exception{
+    public ResponseEntity<?> getProductPostToUpdate(@PathVariable(name="userId") Long userId , @PathVariable(name="productId") Long productId,
+                                                    @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
 
         try {
             ProductResponseDto.EditDto editDto
@@ -158,16 +158,25 @@ public class ProductController {
     //상품상세/ 상품 정보 수정하기
     @PutMapping("/edit/{userId}/{productId}")
     public ResponseEntity<?> updateProductPost(@PathVariable(name="userId") Long userId , @PathVariable(name="productId") Long productId ,
-                                               @RequestBody @Valid ProductPostRequestDto requestDTO, Errors errors) throws Exception{
+                                               @RequestBody @Valid ProductPostRequestDto requestDTO,@RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         productService.updateProductPost(requestDTO,userId,productId);
         return ResponseEntity.ok(HttpStatus.CREATED);
     }
 
-
     //상품 상세/ 상품 판매 상태 변경하기
     //판매완료로 변경
     @PutMapping("/soldOut/{userId}")
-    public ResponseEntity<?> updateProductPostStatusSoldOut(@PathVariable(name="userId") Long userId,@RequestBody @Valid ProductRequestDto requestDto ,Errors errors) throws Exception{
+    public ResponseEntity<?> updateProductPostStatusSoldOut(@PathVariable(name="userId") Long userId,@RequestBody @Valid ProductRequestDto requestDto
+            , @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.updateProductPostStatus(requestDto,userId);
         } catch (IllegalArgumentException e) {
@@ -181,7 +190,13 @@ public class ProductController {
     //상품 상세/ 상품 판매 상태 변경하기
     //판매중으로 변경
     @PutMapping("/onSale/{userId}")
-    public ResponseEntity<?> updateProductPostStatusOnSale(@PathVariable(name="userId") Long userId,@RequestBody @Valid ProductRequestDto requestDto ,Errors errors) throws Exception{
+    public ResponseEntity<?> updateProductPostStatusOnSale(@PathVariable(name="userId") Long userId,@RequestBody @Valid ProductRequestDto requestDto
+            , @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.updateProductPostStatus(requestDto,userId);
         } catch (IllegalArgumentException e) {
@@ -194,10 +209,13 @@ public class ProductController {
 
     //상품 상세/ 상품 글 숨기기 || 숨김 해제하기
     @PutMapping("/private/{userId}/{productId}")
-    public ResponseEntity<?> updateProductPostPrivate(@PathVariable(name="userId") Long userId , @PathVariable(name="productId") Long productId) throws Exception{
+    public ResponseEntity<?> updateProductPostPrivate(@PathVariable(name="userId") Long userId , @PathVariable(name="productId") Long productId,
+                                                      @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
-            System.out.println(userId);
-            System.out.println(productId);
             productService.updateProductPostPrivate(userId,productId);
         } catch (IllegalArgumentException e) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
@@ -209,7 +227,12 @@ public class ProductController {
 
     //상품 상세/ 상품 삭제하기
     @DeleteMapping("/delete/{userId}/{productId}")
-    public ResponseEntity<?> deleteProductPost(@PathVariable(name="userId") Long userId ,@PathVariable(name="productId") Long productId) throws Exception{
+    public ResponseEntity<?> deleteProductPost(@PathVariable(name="userId") Long userId ,@PathVariable(name="productId") Long productId,
+                                               @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.deleteProductPost(userId,productId);
         } catch (IllegalArgumentException e) {
@@ -222,8 +245,12 @@ public class ProductController {
 
     // 상품 찜하기
     @PostMapping("/select/{userId}/{productId}")
-    public ResponseEntity<?> selectProduct(@PathVariable(name="userId") Long userId ,@PathVariable(name="productId") Long productId)
-            throws Exception{
+    public ResponseEntity<?> selectProduct(@PathVariable(name="userId") Long userId ,@PathVariable(name="productId") Long productId,
+                                           @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.selectProduct(userId,productId);
         } catch (IllegalArgumentException e) {
@@ -235,8 +262,12 @@ public class ProductController {
 
     // 상품 찜 해제
     @DeleteMapping("/deselect/{userId}/{productId}")
-    public ResponseEntity<?> deselectProduct(@PathVariable(name="userId") Long userId ,@PathVariable(name="productId") Long productId)
+    public ResponseEntity<?> deselectProduct(@PathVariable(name="userId") Long userId ,@PathVariable(name="productId") Long productId,  @RequestHeader("Authorization") String authorizationHeader)
             throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.deselectProduct(userId,productId);
         } catch (IllegalArgumentException e) {
@@ -250,8 +281,12 @@ public class ProductController {
     // /api/products/block/{userId}/{blockedUserId}
     @PostMapping("/block/{userId}/{blockedUserId}")
     public ResponseEntity<?> blockedUser(@PathVariable(name="userId") Long userId
-            ,@PathVariable(name="blockedUserId") Long blockedUserId)
+            ,@PathVariable(name="blockedUserId") Long blockedUserId, @RequestHeader("Authorization") String authorizationHeader)
             throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.blockedUser(userId,blockedUserId);
         } catch (IllegalArgumentException e) {
@@ -265,7 +300,12 @@ public class ProductController {
     // /api/products/search/save?userId={userId}&searchName={searchName}
     @PostMapping("/search/save")
     public ResponseEntity<?> saveRecentSearchLog(@RequestParam(name="userId") Long userId,
-                                                 @RequestParam(name="searchName") String searchName ) throws Exception{
+                                                 @RequestParam(name="searchName") String searchName ,
+                                                 @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.saveRecentSearchLog(userId,searchName);
         } catch (IllegalArgumentException e) {
@@ -278,8 +318,13 @@ public class ProductController {
     // 최근 검색어 조회
     // /api/products/search/list?userId={userId}
     @GetMapping("/search/list")
-    public ResponseEntity<?> findRecentSearchLogs(@RequestParam(name="userId") Long userId)
+    public ResponseEntity<?> findRecentSearchLogs(@RequestParam(name="userId") Long userId
+            , @RequestHeader("Authorization") String authorizationHeader)
             throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             return ResponseEntity.ok(productService.findRecentSearchLogs(userId));
         } catch (IllegalArgumentException e) {
@@ -292,7 +337,11 @@ public class ProductController {
     // /api/products/search/delete?userId={userId}&searchName={searchName}
     @DeleteMapping("/search/delete")
     public ResponseEntity<?> deleteSearchName(@RequestParam(name="userId") Long userId,
-                                              @RequestParam(name="searchName") String searchName) throws Exception{
+                                              @RequestParam(name="searchName") String searchName, @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             Long count = productService.deleteRecentSearchLog(userId,searchName);
             if (count==1)
@@ -310,7 +359,13 @@ public class ProductController {
     // 최근 검색어 전체 삭제
     // /api/products/search/delete/all?userId={userId}
     @DeleteMapping("/search/delete/all")
-    public ResponseEntity<?> deleteAllSearchNameByUser(@RequestParam(name="userId") Long userId) throws Exception{
+    public ResponseEntity<?> deleteAllSearchNameByUser(@RequestParam(name="userId") Long userId,
+                                                       @RequestHeader("Authorization") String authorizationHeader) throws Exception{
+
+        if(!tokenService.isEqualsUserIdJWT(authorizationHeader,userId))
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("accessToken의 userId 불일치");
+
         try {
             productService.deleteAllSearchNameByUser(userId);
         } catch (IllegalArgumentException e) {
